change double to float
git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench@10249 8b29bb6e-589a-435a-9a12-3ab907be3462

+++ components/aracne-java/src/org/geworkbench/components/aracne/data/AracneInput.java
@@ -106,7 +106,7 @@ public class AracneInput {
 		this.kernelWidth = kernelWidth;
 	}
 
-	public double getKernelWidth() {
+	public float getKernelWidth() {
 		return this.kernelWidth;
 	}
